Move mutation logic to custom hook
